Reduced Technical Debt

+++ src/nz/ac/aut/ense701/gamemodel/occupants/Hazard.java
@@ -45,7 +45,7 @@ public class Hazard extends Occupant {
      */
     public boolean isFatal()
     {
-        return impact == FATAL_IMPACT;
+        return (int) impact == (int)FATAL_IMPACT;
     }
     
     /**
